requirements.txt
pyyaml
pytest
pytest-html
pytest-loop
pytest-repeat
pyserial
pysnmp
uiautomator2
uiautomator
roku
selenium